+++ easypy/signals.py
@@ -146,12 +146,20 @@ class Signal:
 
     ALL = {}
 
+    @staticmethod
+    def get_main_thread():
+        """
+        This function is used to check that new Signal objects are only created from the main thread.
+        If your App's main thread is not threading.main_thread(), you may patch this function.
+        """
+        return threading.main_thread()
+
     def __new__(cls, name, asynchronous=None, swallow_exceptions=False, log=True):
         try:
             return cls.ALL[name]
         except KeyError:
             pass
-        assert threading.main_thread() is threading.current_thread(), "Can only create Signal objects from the MainThread"
+        assert cls.get_main_thread() is threading.current_thread(), "Can only create Signal objects from the MainThread"
         signal = object.__new__(cls)
         # we use this to track the calling of various callbacks. we use this short id so it doesn't overflow the logs
         signal.id = make_id(name)
@@ -376,6 +384,24 @@ run_sync = _set_handler_params(asynchronous=False)
 
 
 def dont_register(func):
+    """
+    Decorator for hiding 'on_...' methods from ``register_object``::
+
+        class MyObject():
+
+            def __init__(self):
+                signals.register_object(self)
+
+            def on_init(self):
+                # this gets registered with the `on_init` signal
+                pass
+
+            @dont_register
+            def on_the_table(self):
+                # this will not get registered
+                pass
+
+    """
     func._signals_dont_register = True
     return func